move feedback form state out of project gallery and into feedback for reusability
