+++ src/components/app/project-gallery/project-gallery.tsx
@@ -1,44 +1,10 @@
-import React, {
-  ChangeEvent,
-  FC,
-  Fragment,
-  SyntheticEvent,
-  useEffect,
-  useState,
-} from 'react';
-import styled from 'styled-components';
+import React, { FC, Fragment, useState, useEffect } from 'react';
 
 import Panel from './panel';
-import {
-  ApiResponse,
-  ErrorResponse,
-  Feedback,
-  Project,
-  ServiceResolver,
-} from '@api';
-import { SecondaryButton } from '@components/shared/buttons';
+import { ApiResponse, ErrorResponse, Project, ServiceResolver } from '@api';
 import { FeedbackForm } from '@components/shared/form';
 import { CloseButton, Ribbon } from '@components/shared/ribbons';
 import { Loader, Wrapper } from '@components/shared';
-import { getNavigatorInfo } from '@utils/browser-utils';
-
-const FeedbackWrapper = styled(Wrapper)`
-  display: flex;
-  justify-content: flex-end;
-  min-height: inherit;
-  padding-bottom: 0;
-  margin-top: 1em;
-
-  @media screen and (max-width: ${({ theme }) => theme.sizes.width.small}) {
-    padding-bottom: 0;
-    padding-top: 0;
-  }
-`;
-
-const FeedbackButton = styled(SecondaryButton)`
-  margin-right: ${({ disabled }) => (disabled ? '1em' : 0)};
-  box-shadow: 1px 1px 2px ${({ theme }) => theme.colors.shadow};
-`;
 
 const ProjectGallery: FC = () => {
   const [projects, setProjects] = useState<Project[]>([]);
@@ -46,42 +12,6 @@ const ProjectGallery: FC = () => {
   const [isLoading, setIsLoading] = useState<boolean>(true);
 
   const [success, setSuccess] = useState<string | null>('');
-  const [feedback, setFeedback] = useState<string>('');
-  const [showFeedbackForm, setShowFeedbackForm] = useState<boolean>(false);
-
-  const handleSendClick = async () => {
-    setError(null);
-
-    const api = ServiceResolver.apiResolver();
-    const navigatorInfo = getNavigatorInfo();
-    let additionalFeedbackInfo = 'Browser info:\n';
-    for (const [key, value] of Object.entries(navigatorInfo)) {
-      additionalFeedbackInfo = additionalFeedbackInfo.concat(
-        `${key}: ${value}\n`,
-      );
-    }
-    const feedbackInfo = `${feedback}\n\n\n\n${additionalFeedbackInfo}`;
-    try {
-      const response = (await api.sendFeedback({
-        content: feedbackInfo,
-      })) as ApiResponse<Feedback | ErrorResponse>;
-
-      if (response.ok) {
-        setShowFeedbackForm(false);
-        setSuccess('Feedback sent successfully');
-        setFeedback('');
-      } else {
-        setError('Failed to send feedback');
-      }
-    } catch (err) {
-      setError('Failed to send feedback');
-    }
-  };
-
-  const handleCancelClick = (e: SyntheticEvent) => {
-    e.preventDefault();
-    setShowFeedbackForm(false);
-  };
 
   useEffect(() => {
     const api = ServiceResolver.apiResolver();
@@ -108,40 +38,9 @@ const ProjectGallery: FC = () => {
     fetchContent();
   }, []);
 
-  const handleDocumentClick = (e: MouseEvent) => {
-    if (!(e.target as HTMLElement).className.includes('feedback')) {
-      setShowFeedbackForm(false);
-      document.removeEventListener('click', handleDocumentClick);
-    }
-  };
-
-  const handleFormChange = (e: ChangeEvent<HTMLInputElement>) =>
-    setFeedback(e.target.value);
-
-  const handleFeedbackButtonOnClick = () => {
-    setShowFeedbackForm(true);
-    document.addEventListener('click', handleDocumentClick);
-  };
-
   return (
     <Fragment>
-      <FeedbackWrapper>
-        {showFeedbackForm && (
-          <FeedbackForm
-            handleSendClick={handleSendClick}
-            handleCancelClick={handleCancelClick}
-            handleChange={handleFormChange}
-            value={feedback}
-          />
-        )}
-
-        <FeedbackButton
-          disabled={showFeedbackForm}
-          onClick={handleFeedbackButtonOnClick}
-        >
-          💡 Got feedback?
-        </FeedbackButton>
-      </FeedbackWrapper>
+      <FeedbackForm />
 
       <Wrapper>
         {success && (
